+++ python/shape_complementarity/__init__.py
@@ -3,7 +3,6 @@ from shape_complementarity.batch import score_many
 from shape_complementarity.io import (
     from_biotite,
     from_boltzgen_refold,
-    from_boltzgen_structure,
     from_pdb,
     from_structure,
 )
@@ -14,7 +13,6 @@ __all__ = [
     "from_pdb",
     "from_structure",
     "from_biotite",
-    "from_boltzgen_structure",
     "from_boltzgen_refold",
     "score_many",
 ]